feat(blog): delete irrelevant BaaS article
This article is not so relevant to me or my blog anymore.
I wrote it not only to learn to write such articles, and present a
technical subject concisely, but also as a test for my development of my
Blogs framework/code for this Astro website of mine.

As I am writing a relevant article, I have removed this article from my
blog.